pyproject.toml
[tool.poetry]
name = "meduzzen-service-app"
version = "0.1.0"
description = ""
authors = ["vitayushchyk <vitay2715@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standard"], version = "^0.115.2" }
uvicorn = "^0.32.0"
pydantic-settings = "^2.6.0"
alembic = "^1.13.3"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.29" }
asyncpg = "^0.30.0"
passlib = "^1.7.4"
pyjwt = "^2.9.0"
authlib = "^1.3.2"
auth0-python = "^4.7.2"
redis = "^5.2.1"
apscheduler = "^3.11.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"